pyproject.toml
[tool.poetry]
name = "sns"
version = "0.1.0"
description = ""
authors = ["haleykimm <accidentalgap@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.1"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.0"
django-cors-headers = "^3.13.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
